==> Final Analitica/Notebooks/1.down_data.ipynb <==
# Descarga de archivos y almacenamiento en 
# carpeta Input que tendra todos los archivos sin procesar

# librerias requeridas para la descarga y creación de archivo csv en crudo
import requests
import os.path
import os

# Url de datos historicos oficiales del Gobierno nacional

url_dataset = 'https://www.datos.gov.co/api/views/gt2j-8ykr/rows.csv?accessType=DOWNLOAD'

# Creación de Directorio para almacenar archivos generados en todas las rutinas

'./'
if os.path.split(os.path.abspath('.'))[-1] == 'src':
    '../Input_data'
# datos de salida disponibles en el directorio "../Output/".
'./'
if os.path.split(os.path.abspath('.'))[-1] == 'src':
    '../Output'

# Crear archivo data_covid.csv y almacenarlo en directorio Input
with requests.get(url_dataset) as dataset_oficial:
    with open(os.path.join('../Input_data', 'data_covid.csv'), 'wb') as dataset_file:
        dataset_file.write(dataset_oficial.content)
