==> Final Analitica/Notebooks/2.Clear_data.ipynb <==
# packages, librerias y dependencias necesarias para el ejercicio de limpieza de datos
import pandas as pd
import numpy as np
import requests
import os.path
import os
import datetime
import dateutil
import subprocess
import sys
import tempfile
!pip install unidecode
import unidecode

# Crear DataFrame para limpieza de datos llamando csv de datos crudos 
dfcovid = pd.read_csv(os.path.join('../Input_data', 'data_covid.csv'))

# nombre de atributo en mayuscula y sin acento utilizando unicode

dfcovid.columns = [unidecode.unidecode(value).upper() for value in dfcovid.columns]

# Eliminar ciudades diferentes a las 5 seleccionadas:
# Medellín 5001
# Bogotá 11001
# Barranquilla 8001
# Cali 76001
# Cartagena 13001

dfcovid=dfcovid.drop(dfcovid[(dfcovid['CODIGO DIVIPOLA'] != 5001)
                             &(dfcovid['CODIGO DIVIPOLA'] != 11001)
                             &(dfcovid['CODIGO DIVIPOLA'] != 8001)
                             &(dfcovid['CODIGO DIVIPOLA'] != 76001)
                             &(dfcovid['CODIGO DIVIPOLA'] != 13001)].index)

#Eliminar columnas que no aportan al desarrollo del proyecto

dfcovid = dfcovid.drop(['NOMBRE GRUPO ETNICO','PERTENENCIA ETNICA','CODIGO PAIS','CODIGO DEPARTAMENTO',], axis=1)

#Verificación ciudades seleccionadas
dfcovid['CIUDAD DE UBICACION'].unique()

#Convertir las columnas a tipo string para un facil manejo
dfcovid = dfcovid.astype({"DEPARTAMENTO O DISTRITO ": str, "TIPO RECUPERACION": str,"PAIS DE PROCEDENCIA": str,"ESTADO": str, "ATENCION": str})

#Transformar los datos de texto en mayusculas 
dfcovid['CIUDAD DE UBICACION'] = [unidecode.unidecode(value).upper() for value in dfcovid['CIUDAD DE UBICACION']]
dfcovid['DEPARTAMENTO O DISTRITO '] = [unidecode.unidecode(value).upper() for value in dfcovid['DEPARTAMENTO O DISTRITO ']]
dfcovid['ATENCION'] = [unidecode.unidecode(value).upper() for value in dfcovid['ATENCION']]
dfcovid['SEXO'] = [unidecode.unidecode(value).upper() for value in dfcovid['SEXO']]
dfcovid['TIPO'] = [unidecode.unidecode(value).upper() for value in dfcovid['TIPO']]
dfcovid['ESTADO'] = [unidecode.unidecode(value).upper() for value in dfcovid['ESTADO']]
dfcovid['PAIS DE PROCEDENCIA'] = [unidecode.unidecode(value).upper() for value in dfcovid['PAIS DE PROCEDENCIA']]
dfcovid['TIPO RECUPERACION'] = [unidecode.unidecode(value).upper() for value in dfcovid['TIPO RECUPERACION']]

#Convertir columnas de Fecha a tipo DATETIME
dfcovid['FECHA DE NOTIFICACION'] = pd.to_datetime(dfcovid['FECHA DE NOTIFICACION'])
dfcovid['FECHA DE MUERTE'] = pd.to_datetime(dfcovid['FECHA DE MUERTE'])
dfcovid['FECHA DIAGNOSTICO'] = pd.to_datetime(dfcovid['FECHA DIAGNOSTICO'])
dfcovid['FECHA RECUPERADO'] = pd.to_datetime(dfcovid['FECHA RECUPERADO'])
dfcovid['FECHA REPORTE WEB'] = pd.to_datetime(dfcovid['FECHA REPORTE WEB'])

# Almacenar DataFrame de limpieza de datos en directorio Output como csv

dfcovid.to_csv(os.path.join('../Output', 'data_final.csv'), index=False)
