==> Final Analitica/Notebooks/3.dataframe_ciudades.ipynb <==
# Librerias necesarias para creación de DataFrames de trabajo
import pandas as pd
import numpy as np
import datetime
import os
import os.path

# Leer archivo procesado de datos csv desde directorio Output

dfcovid_ciudades = pd.read_csv(os.path.join('../Output', 'data_final.csv'))

# Creación archivo de trabajo para Medellín

## Crear DataFrame de Medellín

dfmedellin=dfcovid_ciudades[dfcovid_ciudades['CODIGO DIVIPOLA']==5001]

## Datos Diarios de nuevos casos, muertes y recuperados Medellín

dfmedellin_nuevoscasos = pd.to_datetime(dfmedellin['FECHA REPORTE WEB'])
dfmedellin_nuevoscasos_daily = dfmedellin_nuevoscasos.groupby(dfmedellin_nuevoscasos.dt.floor('d')).size().reset_index(name='CASOS')
dfmedellin_nuevoscasos_daily.columns = ['FECHA', 'CASOS']

dfmedellin_muertes= pd.to_datetime(dfmedellin['FECHA DE MUERTE'])
dfmedellin_muertes_daily=dfmedellin_muertes.groupby(dfmedellin_muertes.dt.floor('d')).size().reset_index(name='FALLECIDOS')
dfmedellin_muertes_daily.columns = ['FECHA', 'FALLECIDOS']

dfmedellin_recu = dfmedellin[dfmedellin['FECHA DE MUERTE'].isna()] #se extrae las fechas de muertes que tambien estan en recuperados
dfmedellin_recu = pd.to_datetime(dfmedellin_recu['FECHA RECUPERADO'])
dfmedellin_recu_daily=dfmedellin_recu.groupby(dfmedellin_recu.dt.floor('d')).size().reset_index(name='RECUPERADOS')
dfmedellin_recu_daily.columns = ['FECHA', 'RECUPERADOS']

#Base de trabajo, Crear columnas con datos acumulados

db_medellin = pd.merge(dfmedellin_nuevoscasos_daily[['FECHA','CASOS']],dfmedellin_recu_daily[['FECHA','RECUPERADOS']],on='FECHA',how='left')
db_medellin = pd.merge(db_medellin[['FECHA','CASOS','RECUPERADOS']],dfmedellin_muertes_daily[['FECHA','FALLECIDOS']],on='FECHA',how='left')
db_medellin = db_medellin.fillna(0)
db_medellin['CASOS_ACUMM'] = db_medellin['CASOS'].cumsum()
db_medellin['RECUPERADOS_ACUMM'] = db_medellin['RECUPERADOS'].cumsum()
db_medellin['FALLECIDOS_ACUMM'] = db_medellin['FALLECIDOS'].cumsum()
db_medellin['ACTIVOS'] =db_medellin['CASOS_ACUMM']-db_medellin['RECUPERADOS_ACUMM']-db_medellin['FALLECIDOS_ACUMM']
db_medellin['ACTIVOS_T1']=db_medellin['ACTIVOS'].shift(1)
db_medellin
# Crear scv en directorio Output
db_medellin.to_csv(os.path.join('../Output', 'data_medellin.csv'), index=False)

# Se repite el proceso para todas las ciudades

# BOGOTA

dfbogota=dfcovid_ciudades[dfcovid_ciudades['CODIGO DIVIPOLA']==11001]

#DATOS DIARIOS
dfbogota_nuevoscasos = pd.to_datetime(dfbogota['FECHA REPORTE WEB'])
dfbogota_nuevoscasos_daily = dfbogota_nuevoscasos.groupby(dfbogota_nuevoscasos.dt.floor('d')).size().reset_index(name='CASOS')
dfbogota_nuevoscasos_daily.columns = ['FECHA', 'CASOS']

dfbogota_muertes= pd.to_datetime(dfbogota['FECHA DE MUERTE'])
dfbogota_muertes_daily=dfbogota_muertes.groupby(dfbogota_muertes.dt.floor('d')).size().reset_index(name='FALLECIDOS')
dfbogota_muertes_daily.columns = ['FECHA', 'FALLECIDOS']

dfbogota_recu = dfbogota[dfbogota['FECHA DE MUERTE'].isna()] #se extrae las fechas de muertes que tambien estan en recuperados
dfbogota_recu = pd.to_datetime(dfbogota_recu['FECHA RECUPERADO'])
dfbogota_recu_daily=dfbogota_recu.groupby(dfbogota_recu.dt.floor('d')).size().reset_index(name='RECUPERADOS')
dfbogota_recu_daily.columns = ['FECHA', 'RECUPERADOS']

#BASE DE TRABAJO
db_bogota = pd.merge(dfbogota_nuevoscasos_daily[['FECHA','CASOS']],dfbogota_recu_daily[['FECHA','RECUPERADOS']],on='FECHA',how='left')
db_bogota = pd.merge(db_bogota[['FECHA','CASOS','RECUPERADOS']],dfbogota_muertes_daily[['FECHA','FALLECIDOS']],on='FECHA',how='left')
db_bogota = db_bogota.fillna(0)
db_bogota['CASOS_ACUMM'] = db_bogota['CASOS'].cumsum()
db_bogota['RECUPERADOS_ACUMM'] = db_bogota['RECUPERADOS'].cumsum()
db_bogota['FALLECIDOS_ACUMM'] = db_bogota['FALLECIDOS'].cumsum()
db_bogota['ACTIVOS'] =db_bogota['CASOS_ACUMM']-db_bogota['RECUPERADOS_ACUMM']-db_bogota['FALLECIDOS_ACUMM']
db_bogota['ACTIVOS_T1']=db_bogota['ACTIVOS'].shift(1)
#db_bogota
db_bogota.to_csv(os.path.join('../Output', 'data_bogota.csv'), index=False)

# CALI

dfcali=dfcovid_ciudades[dfcovid_ciudades['CODIGO DIVIPOLA']==76001]

#DATOS DIARIOS
dfcali_nuevoscasos = pd.to_datetime(dfcali['FECHA REPORTE WEB'])
dfcali_nuevoscasos_daily = dfcali_nuevoscasos.groupby(dfcali_nuevoscasos.dt.floor('d')).size().reset_index(name='CASOS')
dfcali_nuevoscasos_daily.columns = ['FECHA', 'CASOS']

dfcali_muertes= pd.to_datetime(dfcali['FECHA DE MUERTE'])
dfcali_muertes_daily=dfcali_muertes.groupby(dfcali_muertes.dt.floor('d')).size().reset_index(name='FALLECIDOS')
dfcali_muertes_daily.columns = ['FECHA', 'FALLECIDOS']

dfcali_recu = dfcali[dfcali['FECHA DE MUERTE'].isna()] #se extrae las fechas de muertes que tambien estan en recuperados
dfcali_recu = pd.to_datetime(dfcali_recu['FECHA RECUPERADO'])
dfcali_recu_daily=dfcali_recu.groupby(dfcali_recu.dt.floor('d')).size().reset_index(name='RECUPERADOS')
dfcali_recu_daily.columns = ['FECHA', 'RECUPERADOS']

#BASE DE TRABAJO
db_cali = pd.merge(dfcali_nuevoscasos_daily[['FECHA','CASOS']],dfcali_recu_daily[['FECHA','RECUPERADOS']],on='FECHA',how='left')
db_cali = pd.merge(db_cali[['FECHA','CASOS','RECUPERADOS']],dfcali_muertes_daily[['FECHA','FALLECIDOS']],on='FECHA',how='left')
db_cali = db_cali.fillna(0)
db_cali['CASOS_ACUMM'] = db_cali['CASOS'].cumsum()
db_cali['RECUPERADOS_ACUMM'] = db_cali['RECUPERADOS'].cumsum()
db_cali['FALLECIDOS_ACUMM'] = db_cali['FALLECIDOS'].cumsum()
db_cali['ACTIVOS'] =db_cali['CASOS_ACUMM']-db_cali['RECUPERADOS_ACUMM']-db_cali['FALLECIDOS_ACUMM']
db_cali['ACTIVOS_T1']=db_cali['ACTIVOS'].shift(1)
#db_cali
db_cali.to_csv(os.path.join('../Output', 'data_cali.csv'), index=False)

# CARTAGENA

dfcartagena=dfcovid_ciudades[dfcovid_ciudades['CODIGO DIVIPOLA']==13001]

#DATOS DIARIOS
dfcartagena_nuevoscasos = pd.to_datetime(dfcartagena['FECHA REPORTE WEB'])
dfcartagena_nuevoscasos_daily = dfcartagena_nuevoscasos.groupby(dfcartagena_nuevoscasos.dt.floor('d')).size().reset_index(name='CASOS')
dfcartagena_nuevoscasos_daily.columns = ['FECHA', 'CASOS']

dfcartagena_muertes= pd.to_datetime(dfcartagena['FECHA DE MUERTE'])
dfcartagena_muertes_daily=dfcartagena_muertes.groupby(dfcartagena_muertes.dt.floor('d')).size().reset_index(name='FALLECIDOS')
dfcartagena_muertes_daily.columns = ['FECHA', 'FALLECIDOS']

dfcartagena_recu = dfcartagena[dfcartagena['FECHA DE MUERTE'].isna()] #se extrae las fechas de muertes que tambien estan en recuperados
dfcartagena_recu = pd.to_datetime(dfcartagena_recu['FECHA RECUPERADO'])
dfcartagena_recu_daily=dfcartagena_recu.groupby(dfcartagena_recu.dt.floor('d')).size().reset_index(name='RECUPERADOS')
dfcartagena_recu_daily.columns = ['FECHA', 'RECUPERADOS']

#BASE DE TRABAJO
db_cartagena = pd.merge(dfcartagena_nuevoscasos_daily[['FECHA','CASOS']],dfcartagena_recu_daily[['FECHA','RECUPERADOS']],on='FECHA',how='left')
db_cartagena = pd.merge(db_cartagena[['FECHA','CASOS','RECUPERADOS']],dfcartagena_muertes_daily[['FECHA','FALLECIDOS']],on='FECHA',how='left')
db_cartagena = db_cartagena.fillna(0)
db_cartagena['CASOS_ACUMM'] = db_cartagena['CASOS'].cumsum()
db_cartagena['RECUPERADOS_ACUMM'] = db_cartagena['RECUPERADOS'].cumsum()
db_cartagena['FALLECIDOS_ACUMM'] = db_cartagena['FALLECIDOS'].cumsum()
db_cartagena['ACTIVOS'] =db_cartagena['CASOS_ACUMM']-db_cartagena['RECUPERADOS_ACUMM']-db_cartagena['FALLECIDOS_ACUMM']
db_cartagena['ACTIVOS_T1']=db_cartagena['ACTIVOS'].shift(1)
#db_cartagena
db_cartagena.to_csv(os.path.join('../Output', 'data_cartagena.csv'), index=False)

# BARRANQUILLA

dfbarranquilla=dfcovid_ciudades[dfcovid_ciudades['CODIGO DIVIPOLA']==8001]

#DATOS DIARIOS
dfbarranquilla_nuevoscasos = pd.to_datetime(dfbarranquilla['FECHA REPORTE WEB'])
dfbarranquilla_nuevoscasos_daily = dfbarranquilla_nuevoscasos.groupby(dfbarranquilla_nuevoscasos.dt.floor('d')).size().reset_index(name='CASOS')
dfbarranquilla_nuevoscasos_daily.columns = ['FECHA', 'CASOS']

dfbarranquilla_muertes= pd.to_datetime(dfbarranquilla['FECHA DE MUERTE'])
dfbarranquilla_muertes_daily=dfbarranquilla_muertes.groupby(dfbarranquilla_muertes.dt.floor('d')).size().reset_index(name='FALLECIDOS')
dfbarranquilla_muertes_daily.columns = ['FECHA', 'FALLECIDOS']

dfbarranquilla_recu = dfbarranquilla[dfbarranquilla['FECHA DE MUERTE'].isna()] #se extrae las fechas de muertes que tambien estan en recuperados
dfbarranquilla_recu = pd.to_datetime(dfbarranquilla_recu['FECHA RECUPERADO'])
dfbarranquilla_recu_daily=dfbarranquilla_recu.groupby(dfbarranquilla_recu.dt.floor('d')).size().reset_index(name='RECUPERADOS')
dfbarranquilla_recu_daily.columns = ['FECHA', 'RECUPERADOS']

#BASE DE TRABAJO
db_barranquilla = pd.merge(dfbarranquilla_nuevoscasos_daily[['FECHA','CASOS']],dfbarranquilla_recu_daily[['FECHA','RECUPERADOS']],on='FECHA',how='left')
db_barranquilla = pd.merge(db_barranquilla[['FECHA','CASOS','RECUPERADOS']],dfbarranquilla_muertes_daily[['FECHA','FALLECIDOS']],on='FECHA',how='left')
db_barranquilla = db_barranquilla.fillna(0)
db_barranquilla['CASOS_ACUMM'] = db_barranquilla['CASOS'].cumsum()
db_barranquilla['RECUPERADOS_ACUMM'] = db_barranquilla['RECUPERADOS'].cumsum()
db_barranquilla['FALLECIDOS_ACUMM'] = db_barranquilla['FALLECIDOS'].cumsum()
db_barranquilla['ACTIVOS'] =db_barranquilla['CASOS_ACUMM']-db_barranquilla['RECUPERADOS_ACUMM']-db_barranquilla['FALLECIDOS_ACUMM']
db_barranquilla['ACTIVOS_T1']=db_barranquilla['ACTIVOS'].shift(1)
#db_barranquilla
db_barranquilla.to_csv(os.path.join('../Output', 'data_barranquilla.csv'), index=False)
